requirements/test.txt
coverage>=4.3.4
mock>=2.0.0
flake8>=3.3.0
pytest==3.0.7
pytest-django==3.1.2
django_dynamic_fixture==1.9.4
Faker==0.8.6
vcrpy==1.11.1
pytest-mock==1.6.3
